:truck: Moved Greengraph exercises to module 02
