+++ docs/_toc.yml
@@ -30,11 +30,12 @@ parts:
     - file: module02_intermediate_python/02_00_comprehensions.ipynb
     - file: module02_intermediate_python/02_01_functions.ipynb
     - file: module02_intermediate_python/02_02_using_libraries.ipynb
-    - file: module02_intermediate_python/02_03_defining_classes.ipynb
-    - file: module02_intermediate_python/02_04_working_with_files.ipynb
-    - file: module02_intermediate_python/02_05_getting_data_from_the_internet.ipynb
-    - file: module02_intermediate_python/02_06_bringing_it_together.ipynb
-    - file: module02_intermediate_python/02_07_classroom_exercises.ipynb
+    - file: module02_intermediate_python/02_03_working_with_files.ipynb
+    - file: module02_intermediate_python/02_04_getting_data_from_the_internet.ipynb
+    - file: module02_intermediate_python/02_05_bringing_it_together.ipynb
+    - file: module02_intermediate_python/02_06_defining_classes.ipynb
+    - file: module02_intermediate_python/02_07_understanding_greengraph.ipynb
+    - file: module02_intermediate_python/02_08_classroom_exercises.ipynb
   - file: module03_research_data_in_python/index.md
     sections:
     - file: module03_research_data_in_python/03_00_fields_and_records.ipynb
@@ -42,9 +43,8 @@ parts:
     - file: module03_research_data_in_python/03_02_plotting_with_matplotlib.ipynb
     - file: module03_research_data_in_python/03_03_NumPy.ipynb
     - file: module03_research_data_in_python/03_04_NumPy_continued.ipynb
-    - file: module03_research_data_in_python/03_05_understanding_greengraph.ipynb
-    - file: module03_research_data_in_python/03_06_boids.ipynb
-    - file: module03_research_data_in_python/03_07_classroom_exercises.ipynb
+    - file: module03_research_data_in_python/03_05_boids.ipynb
+    - file: module03_research_data_in_python/03_06_classroom_exercises.ipynb
   - file: module04_version_control_with_git/index.md
     sections:
     - file: module04_version_control_with_git/04_00_introduction.ipynb
