+++ module02_intermediate_python/index.md
@@ -13,12 +13,13 @@
 - [2.0 Comprehensions](02_00_comprehensions.ipynb) (10 minutes)
 - [2.1 Functions](02_01_functions.ipynb) (15 minutes)
 - [2.2 Using Libraries](02_02_using_libraries.ipynb) (5 minutes)
-- [2.3 Defining your own classes](02_03_defining_classes.ipynb) (20 minutes)
-- [2.4 Working with files](02_04_working_with_files.ipynb) (15 minutes)
-- [2.5 Getting data from the Internet](02_05_getting_data_from_the_internet.ipynb) (10 minutes)
-- [2.6 Bringing it all together](02_06_bringing_it_together.ipynb) (20 minutes)
+- [2.3 Working with files](02_03_working_with_files.ipynb) (15 minutes)
+- [2.4 Getting data from the Internet](02_04_getting_data_from_the_internet.ipynb) (10 minutes)
+- [2.5 Bringing it all together](02_05_bringing_it_together.ipynb) (20 minutes)
+- [2.6 Defining your own classes](02_06_defining_classes.ipynb) (20 minutes)
+- [2.7 Writing a Greengraph class](02_07_writing_a_greengraph_class.ipynb) (10 minutes)
 
-**Total time:** 1 hr 35 minutes
+**Total time:** 1 hr 45 minutes
 
 ## Exercises
