--- module03_research_data_in_python/greengraph/__init__.py
@@ -1,1 +1,0 @@
-from .graph import Greengraph
